+++ packages/qwik/src/validation/minTotalSize.ts
@@ -1,7 +1,7 @@
 import { $, type NoSerialize, type QRL } from '@builder.io/qwik';
 import type { MaybeValue } from '../types';
 
-type Value = MaybeValue<NoSerialize<Blob[]> | NoSerialize<File[]>>;
+type Value = MaybeValue<NoSerialize<Blob>[] | NoSerialize<File>[]>;
 
 /**
  * Creates a validation functions that validates total file size of a file list.
@@ -17,7 +17,7 @@ export function minTotalSize(
 ): QRL<(value: Value) => string> {
   return $((value: Value) =>
     value?.length &&
-    [...value].reduce((size, file) => size + file.size, 0) < requirement
+    [...value].reduce((size, file) => size + file!.size, 0) < requirement
       ? error
       : ''
   );